Inline the schema
